# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumo Task API",
    "version": "1.0.0",
    "description": "REST API for Lumo Task — an Eisenhower-matrix task manager.\n\n**This document is generated** from the route registry in `@lumo/contracts` (`src/registry.ts`) and the Zod schemas it references. It is never hand-edited, and a parity test fails the build if it stops matching the routes the server actually mounts.\n\n**Authentication:** every `/v1/*` route requires a Bearer token except the auth and OAuth handshake endpoints. Call `POST /v1/auth/signin` to obtain one, then click **Authorize** and paste it.\n\n**Errors** all share one envelope: `{ error: { code, message, fields? } }`."
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuthSession": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "recoveryCode": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "nullable": true
              },
              "name": {
                "type": "string"
              },
              "initials": {
                "type": "string"
              },
              "local": {
                "type": "boolean"
              },
              "emailVerified": {
                "type": "boolean"
              },
              "plan": {
                "type": "string"
              },
              "renewsAt": {
                "type": "string",
                "nullable": true
              },
              "stats": {
                "type": "object",
                "properties": {
                  "tasks": {
                    "type": "number"
                  },
                  "pomodoros": {
                    "type": "number"
                  },
                  "syncOK": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "tasks",
                  "pomodoros",
                  "syncOK"
                ]
              }
            },
            "required": [
              "id",
              "username",
              "email",
              "name",
              "initials",
              "local",
              "emailVerified",
              "plan",
              "renewsAt",
              "stats"
            ]
          }
        },
        "required": [
          "token",
          "refreshToken",
          "user"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/auth/github/config": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Is GitHub sign-in available?",
        "operationId": "getV1AuthGithubConfig",
        "description": "Lets the client hide the button when the deployment has no GitHub app configured.",
        "responses": {
          "200": {
            "description": "Availability flag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "enabled"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/github/start": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Begin the GitHub OAuth handshake",
        "operationId": "getV1AuthGithubStart",
        "responses": {
          "302": {
            "description": "Redirect to GitHub's consent screen"
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/github/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GitHub OAuth callback",
        "operationId": "getV1AuthGithubCallback",
        "description": "Validates the state parameter, then redirects back into the app with a one-time code.",
        "responses": {
          "302": {
            "description": "Redirect into the app"
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/github/exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange the one-time OAuth code for a session",
        "operationId": "postV1AuthGithubExchange",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — see `error.fields` for the offending inputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Code unknown, expired, or already used",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/github/link": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Link a GitHub identity to the signed-in account",
        "operationId": "postV1AuthGithubLink",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "GitHub identity linked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "That GitHub identity is already linked elsewhere",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}